Add support for "signed" registers (#113)
* ShineWiFi-ModBus.ino: Rename JSONString to JSONChars
It is conflicting with ArduinoJSON namespace and also
a bit confusing name. To resolve conflict we need to
rename it.

* Add support for "signed" registers
As SPF protocol shows, battery balance are shown as "signed"
value, e.g. if we have value -1000 it means battery being charged,
if it is 1000 it means battery is discharged.
(Holding register 77, 32-bit value, SPF5000 Modbus protocol)
Tested only on 32-bit input register.

Signed-off-by: Denys Fedoryshchenko <denys.f@collabora.com>