feat: add hostname and static ip settings
this adds support for setting a custom hostname and static ip settings.
If more than once inverter is in the installation it is good to be able
to distinguish the different inverters by name.
For sensors it is a good practice to give static ip addresses to them
especially if they are polled.

Signed-off-by: Peter Lieven <pl@dlhnet.de>