fix: handle state of red led correctly
the red led is used to indicate different error scenarious.
Currently it is used for failed mqtt connection and failed mobdus readout.
Each of these errors clear and set the led state individually which is not
correct.

This patch changes the behavior to set it if any of the error cases is true
and clears it only if all are false.

While at it also add the unknown shine wifi stick case as an error.

Signed-off-by: Peter Lieven <pl@dlhnet.de>