fix(Modbus): reduce modbus timeout from 2000 to 250ms for 115200 baud
it turns out that a significant portion of modbus requests fails returning in
2000ms stalls of the ESP since the default timeout in the ModbusMaster library
is hardcoded to 2000ms.

In fact I think that the timeouts are busy response messages which
the ModbusMaster can't handle.

Since the ModbusMaster library has not been updated for 8 years with
many PRs for adding a timeout setting change to a fork which basically is
the original library plus the support for setting a timeout.

Since 115200 baud is more than 10x faster than 9600 baud its resonable
to reduce the timeout from 2000ms to 250ms. In my tests the failure rate
was not changed by this reduction.

Signed-off-by: Peter Lieven <pl@dlhnet.de>