feat(sendMetrics): improve create and transfer speed
this change improves the overall generate and transmit time for arround 100 metrics on
my ShineWifiX stick from approx. 80 ms to 50 ms.

The current implementation is slower for 3 reasons:
 - StringStream is not able to reserve memory. Thus each concat will cause realloc which even might leed to heap fragmentation.
 - The printf("%g") in Growatt::metricsAddValue seems to be an expensive operation.
 - Sending slices from a StringStream to WifiClient involves copying data.

While at it also add support for sending a 503 error if the last modbus readout failed or has not yet taken places. This avoids
that zero or outdated data is imported to prometheus.

Signed-off-by: Peter Lieven <pl@dlhnet.de>