feat: add /value endpoint
this adds support for accessing individual register values via http without the need to request the full json or metrics output and parse it.

For example the output power can be queried via:
  http://IP/value/OutputPower

I use it for accessing the inverter data from inside evcc for example which issues a new request for each value it wants to get.

Signed-off-by: Peter Lieven <pl@dlhnet.de>