refactor: remove WifiManager from global context
this moves the WifiManager from global to setup() context. This on one side saves about 500 bytes of heap. On the other side it gives the possibility to enter the config AP also by holding the AP button on power up if for whatever reason the main loop has an issue.

Signed-off-by: Peter Lieven <pl@dlhnet.de>