feat: add support for TL-X(H) inverters
according to the modbus v1.24 specification from 2020-04-28 page 3 the following
inverters must use different input registers (3000+):

- TL-X/TL-XH MIN Type (it seems that they are also called MIC today).
- MOD TL3-XH (this also seems to apply for MID KTL3-XH).

To choose the high input and holding registers set GROWAT_MODBUS_VERION 3000
in Config.h.

The inverters seem to be somewhat compatible to the low register range, but if
a battery system is used the battery system is not reported correctly and holding
registers do not work at all.

This patch also adds support for setting charge and discharge power rates and stop SOCs
as well as AC charging. Currently not supported are getting and setting timeslots
for different priority modes.

Signed-off-by: Peter Lieven <pl@dlhnet.de>