feat: add fragment retry logic to modbus poller
currently if a modbus request fails the logic starts over from the first
input regiter fragment.
This is adds a logic that restarts at the last failed fragment.
About 8-10% of all modbus requests fail due to busy conditions
which the modbus library can't handle.
So retrying only the failed fragment is a big improvement.

Signed-off-by: Peter Lieven <pl@dlhnet.de>